pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "veco"
version = "0.1.0"
description = "Vectorization & RAG Toolkit"
readme = "README.md"
requires-python = ">=3.10,<3.12"
license = "CC0-1.0"
authors = [{ name = "Matthias Pohl" }]
keywords = ["rag", "vector database", "embeddings", "multimodal", "process mining"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Multimedia :: Sound/Audio :: Speech",
]

# Core dependencies (NumPy 2.2.6 + compatible pins)
dependencies = [
  # Numerics & core ML
  "numpy>=1.24,<2.0",
  "scipy>=1.10,<1.12",
  "torch>=2.1,<2.3",
  "torchaudio>=2.1,<2.3",
  "torchvision>=0.16,<0.18",
  "sentence-transformers>=2.2,<3.0",

  # Vector DB & media
  "faiss-cpu>=1.7.4,<1.8",
  "moviepy>=1.0.3,<2.0",

  # NLP / STT / RAG
  "openai-whisper>=20231117,<20250000",
  "ollama>=0.1.6,<0.3",

  # Audio / DSP
  "librosa>=0.10,<0.11",
  "soundfile>=0.12,<0.13",
  "webrtcvad-wheels>=2.0.10,<2.1",

  # Vision / OCR / imaging
  "pytesseract>=0.3.10,<0.4",
  "Pillow>=10.2,<11",

  # Documents
  "pdfplumber>=0.11.0,<0.12",
  "python-docx>=0.8.11,<0.9",
  "python-pptx>=0.6.23,<0.7",

  # Misc
  "speechbrain>=0.5.15,<0.6",
  "tqdm>=4.66,<4.67",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "black",
  "ruff",
  "mypy",
]

pic-describe = [
  "transformers>=4.37,<5",
  "sentencepiece>=0.1.99,<0.2",
  "sacremoses>=0.0.53,<0.1",
]

[project.urls]
Homepage = "https://github.com/MatthiasPohlAmberg/VeCo"
Source = "https://github.com/MatthiasPohlAmberg/VeCo"
Issues = "https://github.com/MatthiasPohlAmberg/VeCo/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["VeCo*"]
